Splits AAPBResults to own component
